+++ NEWS.md
@@ -24,6 +24,7 @@
 * added `prism_data_get_md()` to parse .info.txt metadata, by converting an existing internal function. (#88)
 * `del_early_prov()` now invisibly returns the folders that it removes.
 * `prism_data_clean()` replaces `del_early_prov()` and works with all time steps. It also prompts user to select which folders will be removed before removing them (when R is in interactive mode). (#89)
+* `prism_image()` invisibly returns the `gg` object it creates.
 
 # prism 0.1.0